implemented postpolicy for delete action

+++ SayHiLaravel/app/Http/Controllers/PostController.php
@@ -27,10 +27,7 @@ class PostController extends Controller
     }
     public function destroy(Post $post)
     {
-        if(!$post->OwnedBy(auth()->user()))
-        {
-            dd('You cant delete another user\'s post');
-        }
+        $this->authorize('delete',$post);
         $post->delete();
         return back();
     }
